startup-validator/python_core/requirements.txt
crewai
crewai-tools
langchain-google-genai
langchain-community
streamlit
tavily-python
python-dotenv
